server/requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
openai>=1.50.0
python-dotenv>=1.0.0
python-multipart>=0.0.12
browser-use>=0.12.0
agentmail
google-auth
google-auth-oauthlib
google-api-python-client
